% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VitDpackage.R
\docType{package}
\name{VitDpackage}
\alias{VitDpackage}
\alias{VitDpackage-package}
\title{VitDpackage: A package to store data and reports.}
\description{
The VitDpackage package is designed to store data and reports Dr. Baggerly generated for his CDC talk.
}
\section{VitDpackage functions}{

No function
}

